fix(logs): define BoundLogger in log module
